pyproject.toml
[project]
name = "fastfractal"
version = "0.1.0"
description = "Fast fractal image compression packages"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["numpy>=1.26", "pillow>=10.0"]

[project.optional-dependencies]
dev = ["pytest>=8.0", "mypy>=1.8", "ruff>=0.6", "black>=24.0", "types-Pillow"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[build-system]
requires = ["setuptools>=68", "wheel", "numpy>=1.26"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["fastfractal"]